feat: create all models in separate files
