+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/Key.kt
@@ -1,0 +1,7 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+data class Key(
+    var keyId: String,
+    var key: ByteArray,
+    val type: KeyType
+)
