+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/KeyType.kt
@@ -1,0 +1,6 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+enum class KeyType {
+    LOCAL,
+    REMOTE
+}
