+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/SdpWithTypeModel.kt
@@ -1,0 +1,9 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+/**
+ * kotlin equivalent of privmx::endpoint::stream::SdpWithTypeModel
+ */
+data class SdpWithTypeModel(
+    val sdp: String,
+    val type: String,
+)
