+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/StreamEncryptionMode.kt
@@ -1,0 +1,6 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+enum class StreamEncryptionMode {
+    SINGLE_KEY,
+    MULTIPLE_KEY,
+}
