+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/StreamHandle.kt
@@ -1,0 +1,3 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+value class StreamHandle internal constructor(val value: Long)
