+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/StreamInfo.kt
@@ -1,0 +1,10 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+data class StreamInfo(
+    val id: Long,
+    val userId: String,
+    val metadata: String?,
+    val dummy: Boolean?,
+    val tracks: List<StreamTrackInfo>,
+    val talking: Boolean?,
+)
