+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/StreamPublishResult.kt
@@ -1,0 +1,8 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+import com.simplito.kotlin.privmx_endpoint.model.stream.events.PublishedStreamData
+
+data class StreamPublishResult(
+    val published: Boolean,
+    val data: PublishedStreamData?,
+)
