+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/StreamTrackModification.kt
@@ -1,0 +1,6 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+data class StreamTrackModification(
+    val streamId: Long,
+    val tracks: List<StreamTrackModificationPair>,
+)
