+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/StreamTrackModificationPair.kt
@@ -1,0 +1,6 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+data class StreamTrackModificationPair(
+    val before: StreamTrackInfo?,
+    val after: StreamTrackInfo?,
+)
