+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/TurnCredentials.kt
@@ -1,0 +1,11 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream
+
+/**
+ * kotlin equivalent of privmx::endpoint::stream::TurnCredentials
+ */
+ data class TurnCredentials(
+     val url: String,
+     val username: String,
+     val password: String,
+     val expirationTime: Long
+ )
