+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/NewStreams.kt
@@ -1,0 +1,9 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+
+import com.simplito.kotlin.privmx_endpoint.model.stream.StreamInfo
+
+data class NewStreams(
+    val room: String,
+    val streams: List<StreamInfo>
+)
