+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/StreamEventData.kt
@@ -1,0 +1,7 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+data class StreamEventData(
+    val streamRoomId: String,
+    val streamIds: List<Long>,
+    val userId: String?
+)
