+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/StreamLeftEventData.kt
@@ -1,0 +1,7 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+data class StreamLeftEventData(
+    val streamRoomId: String,
+    val streamId: Long,
+    val userId: String
+)
