+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/StreamPublishedEventData.kt
@@ -1,0 +1,11 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+
+import com.simplito.kotlin.privmx_endpoint.model.stream.StreamInfo
+
+data class StreamPublishedEventData(
+    val streamRoomId: String,
+    val stream:
+    StreamInfo,
+    val userId: String
+)
