+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/StreamRoomDeletedEventData.kt
@@ -1,0 +1,5 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+data class StreamRoomDeletedEventData(
+    val streamRoomId: String
+)
