+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/StreamUnpublishedEventData.kt
@@ -1,0 +1,6 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+data class StreamUnpublishedEventData(
+    val streamRoomId: String,
+    val streamId: Long
+)
