+++ privmx-endpoint/src/commonMain/kotlin/com/simplito/kotlin/privmx_endpoint/model/stream/events/StreamsUpdatedData.kt
@@ -1,0 +1,6 @@
+package com.simplito.kotlin.privmx_endpoint.model.stream.events
+
+data class StreamsUpdatedData(
+    val room: String,
+    val streams: List<UpdatedStreamData>,
+)
